Some colony info on planet list

+++ FrEee.WinForms/Forms/GameForm.cs
@@ -1168,7 +1168,7 @@ namespace FrEee.WinForms.Forms
 		{
 			var budget = Empire.Current.NetIncome[Resource.Research] + Empire.Current.BonusResearch;
 			var forQueue = 100 - Empire.Current.ResearchSpending.Sum(kvp => kvp.Value);
-			return t.Spending.Value * budget / 100 + (Empire.Current.ResearchQueue.FirstOrDefault() == t ? forQueue : 0);
+			return (int)(t.Spending.Value * budget / 100 + (Empire.Current.ResearchQueue.FirstOrDefault() == t ? forQueue : 0));
 		}
 
 		public void ShowResearchForm()
